HTML file that matches open script now opens in browser.
If the current file that is open in the editor matches a file found in "reports/mutation/html/**/" then that file is opened in the user's default browser. If no match is found then the default entry point for reports (i.e. "reports/mutation/html/index.html") is opened instead.